Criado mais testes e também o controller do CreateMeetup

+++ Servico-de-agendamento-do-meetups/src/main/java/com/womkarescode/microservicemeetup/service/impl/CreateMeetupServiceImpl.java
@@ -1,5 +1,6 @@
 package com.womkarescode.microservicemeetup.service.impl;
 
+import com.womkarescode.microservicemeetup.exception.BusinessException;
 import com.womkarescode.microservicemeetup.model.entity.CreateMeetup;
 import com.womkarescode.microservicemeetup.repository.CreateMeetupRepository;
 import com.womkarescode.microservicemeetup.service.CreateMeetupService;
@@ -22,6 +23,9 @@ public class CreateMeetupServiceImpl implements CreateMeetupService {
 
     @Override
     public CreateMeetup saveNewEventMeetup(CreateMeetup createEvent) {
+        if(repository.findByEvent(createEvent.getEvent()).isPresent()){
+            throw new BusinessException("Event already created");
+        }
         return repository.save(createEvent);
     }
 
@@ -37,12 +41,18 @@ public class CreateMeetupServiceImpl implements CreateMeetupService {
 
     @Override
     public CreateMeetup updateEventMeetup(CreateMeetup event) {
+        if(event == null || event.getId() == null){
+            throw new IllegalArgumentException("Event id must not be null");
+        }
         return repository.save(event);
     }
 
     @Override
     public void deleteEventMeetup(CreateMeetup eventMeetup) {
-        repository.delete(eventMeetup);
+        if(repository.findByEvent(eventMeetup.getEvent()).isEmpty()){
+            throw new IllegalArgumentException("Event id must exist");
+        }
+         repository.delete(eventMeetup);
     }
 
     @Override
